# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kitchen Pantry CRM API
  description: |
    A comprehensive RESTful API for managing food service industry relationships, interactions, and sales opportunities. 
    Built on Node.js/Express with Supabase backend integration, supporting iPad-first CRM operations.
  version: 1.0.0
  contact:
    name: Kitchen Pantry CRM Support
    email: support@kitchenpantry.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.kitchenpantry.com/v1
    description: Production server
  - url: https://staging-api.kitchenpantry.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /organizations/{org_id}/contacts:
    get:
      summary: List contacts for organization
      description: Retrieves all contacts within a specific organization with role-based filtering.
      tags:
        - Contacts
      parameters:
        - name: org_id
          in: path
          required: true
          description: Organization ID
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of records per page
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: search
          in: query
          description: Search across contact names, job titles, and email addresses
          schema:
            type: string
        - name: is_primary_contact
          in: query
          description: Filter for primary contacts only
          schema:
            type: boolean
        - name: is_decision_maker
          in: query
          description: Filter for decision makers only
          schema:
            type: boolean
        - name: authority_level
          in: query
          description: Filter by authority level
          schema:
            type: integer
            minimum: 1
            maximum: 5
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /contacts:
    get:
      summary: List contacts
      description: Retrieves contacts accessible to the authenticated user with filtering and search capabilities.
      tags:
        - Contacts
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of records per page
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: organization_id
          in: query
          description: Filter contacts by organization
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          description: Search across contact names, job titles, and email addresses
          schema:
            type: string
        - name: is_primary_contact
          in: query
          description: Filter for primary contacts only
          schema:
            type: boolean
        - name: is_decision_maker
          in: query
          description: Filter for decision makers only
          schema:
            type: boolean
        - name: authority_level
          in: query
          description: Filter by authority level
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: last_interaction_days
          in: query
          description: Filter contacts with interactions within specified days
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create contact
      description: Creates a new contact with organization association and validation.
      tags:
        - Contacts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreateRequest'
      responses:
        '201':
          description: Contact created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /contacts/{id}:
    get:
      summary: Get contact by ID
      description: Retrieves detailed information for a specific contact.
      tags:
        - Contacts
      parameters:
        - name: id
          in: path
          required: true
          description: Contact ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update contact
      description: Updates an existing contact with complete resource replacement.
      tags:
        - Contacts
      parameters:
        - name: id
          in: path
          required: true
          description: Contact ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactUpdateRequest'
      responses:
        '200':
          description: Contact updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      summary: Partially update contact
      description: Performs partial updates to a contact, updating only specified fields.
      tags:
        - Contacts
      parameters:
        - name: id
          in: path
          required: true
          description: Contact ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactPatchRequest'
      responses:
        '200':
          description: Contact updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete contact
      description: Soft deletes a contact by setting is_deleted = true.
      tags:
        - Contacts
      parameters:
        - name: id
          in: path
          required: true
          description: Contact ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Contact deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued by Supabase Auth
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 150
        pages:
          type: integer
          example: 3
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
        next_cursor:
          type: string
          nullable: true
        prev_cursor:
          type: string
          nullable: true
    ResponseMetadata:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Organization schemas
...
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        full_name:
          type: string
        job_title:
          type: string
        department:
          type: string
        is_primary_contact:
          type: boolean
        is_decision_maker:
          type: boolean
        authority_level:
          type: integer
          minimum: 1
          maximum: 5
        email_primary:
          type: string
          format: email
        phone_primary:
          type: string
        mobile_phone:
          type: string
        preferred_contact_method:
          type: string
          enum: [email, phone, text, linkedin]
        preferred_contact_time:
          type: string
        linkedin_url:
          type: string
          format: uri
        communication_preferences:
          type: object
          properties:
            email_frequency:
              type: string
              enum: [daily, weekly, monthly, quarterly]
            content_types:
              type: array
              items:
                type: string
            opt_out_marketing:
              type: boolean
        last_interaction_date:
          type: string
          format: date-time
        next_follow_up_date:
          type: string
          format: date-time
        relationship_score:
          type: integer
          minimum: 1
          maximum: 10
        influence_score:
          type: integer
          minimum: 1
          maximum: 10
        engagement_level:
          type: string
          enum: [low, medium, high]
        interaction_count:
          type: integer
        organization:
          type: object
          properties:
            name:
              type: string
            industry_segment:
              type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ContactCreateRequest:
      type: object
      properties:
        organization_id:
          type: string
          format: uuid
        first_name:
          type: string
          minLength: 1
          maxLength: 100
        last_name:
          type: string
          minLength: 1
          maxLength: 100
        job_title:
          type: string
        department:
          type: string
        is_primary_contact:
          type: boolean
          default: false
        is_decision_maker:
          type: boolean
          default: false
        authority_level:
          type: integer
          minimum: 1
          maximum: 5
        email_primary:
          type: string
          format: email
        phone_primary:
          type: string
        mobile_phone:
          type: string
        preferred_contact_method:
          type: string
          enum: [email, phone, text, linkedin]
        preferred_contact_time:
          type: string
        linkedin_url:
          type: string
          format: uri
        communication_preferences:
          type: object
          properties:
            email_frequency:
              type: string
              enum: [daily, weekly, monthly, quarterly]
            content_types:
              type: array
              items:
                type: string
            opt_out_marketing:
              type: boolean
              default: false
            best_contact_time:
              type: string
      required:
        - organization_id
        - first_name
        - last_name
        - email_primary
    ContactUpdateRequest:
      $ref: '#/components/schemas/ContactCreateRequest'
    ContactPatchRequest:
      type: object
      properties:
        first_name:
          type: string
          minLength: 1
          maxLength: 100
        last_name:
          type: string
          minLength: 1
          maxLength: 100
        job_title:
          type: string
        department:
          type: string
        is_primary_contact:
          type: boolean
        is_decision_maker:
          type: boolean
        authority_level:
          type: integer
          minimum: 1
          maximum: 5
        email_primary:
          type: string
          format: email
        phone_primary:
          type: string
        mobile_phone:
          type: string
        preferred_contact_method:
          type: string
          enum: [email, phone, text, linkedin]
        preferred_contact_time:
          type: string
        linkedin_url:
          type: string
          format: uri
        communication_preferences:
          type: object
          properties:
            email_frequency:
              type: string
              enum: [daily, weekly, monthly, quarterly]
            content_types:
              type: array
              items:
                type: string
            opt_out_marketing:
              type: boolean
            best_contact_time:
              type: string
    ContactResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Contact'
        meta:
          $ref: '#/components/schemas/ResponseMetadata'
    ContactDetailResponse:
      $ref: '#/components/schemas/ContactResponse'
    ContactListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        meta:
          $ref: '#/components/schemas/ResponseMetadata'

    # Interaction schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
              field:
                type: string
              details:
                type: string
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
            retry_after:
              type: integer
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: UNAUTHORIZED
                message: Authentication required
                details: Valid JWT token required in Authorization header
    ForbiddenError:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: FORBIDDEN
                message: Access denied
                details: Insufficient permissions for this resource
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: NOT_FOUND
                message: Resource not found
                details: The requested resource does not exist
    BadRequestError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: BAD_REQUEST
                message: Invalid request format
                details: Request body must be valid JSON
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: CONFLICT
                message: Resource conflict
                details: Organization name already exists
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: VALIDATION_ERROR
                message: Invalid input data
                field: email_primary
                details: Email format is invalid
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: INTERNAL_ERROR
                message: Internal server error
                details: An unexpected error occurred
...